requirements.txt
streamlit
python-dotenv
langchain
langchain-google-genai
langchain-community
langchain-huggingface
sentence-transformers
faiss-cpu
PyMuPDF
